@@ -9,13 +9,12 @@
         "email": "nticaric@gmail.com"
     }],
     "require": {
-        "php": ">=5.3.0",
-        "nesbot/carbon": "~1.0"
+        "php": ">=7.3.0",
+        "nesbot/carbon": "^2.0"
     },
     "require-dev": {
-        "mockery/mockery": "~0.9",
-        "phpunit/phpunit": "~4.0",
-        "codeclimate/php-test-reporter": "dev-master"
+        "mockery/mockery": "^1.0",
+        "phpunit/phpunit": "^9.0"
     },
     "autoload": {
         "psr-4": {
