@@ -1,8 +1,9 @@
 <?php
 
 use Nticaric\Fiskalizacija\Bill\Refund;
+use PHPUnit\Framework\TestCase;
 
-class RefundTest extends \PHPUnit_Framework_TestCase
+class RefundTest extends TestCase
 {
     public function testRefundClass()
     {
